Fixes part of #10415: Refactoring story validators (#10973)
